django_app/requirements.txt
django==2.2
djangorestframework==3.9.2
rest_framework_dyn_serializer==1.3.1
tipsi_tools==1.41.0
